vgt: add one unknow median command handler
we notice there is unknown media command 680b0000 when running
glxgear. although we did not find it in spec, from dumping the
batch buffer, it is very likely its length is 1. so it can be
handled as noop.

Note that this is temporarily solution. need find this instruction
in spec later.

Signed-off-by: Yu Ke <ke.yu@intel.com>