vGT: use MI_STORE_DATA_IMM to signal the end of the context switch
earlier we simply compare the ring head and tail, which doesn't mean
that all the commands have been retired from the GPU. Put a data
store command at the end of the context switch, and then poll the
status of the store destination. This way can ensure the context
switch really ends.

One potential issue still remains, that we now utilize the guest
ring buffer for submitting our context switch commands. This is
problematic in the SMP environment since the guest may write to that
ring buffer in the same time. This will be fixed later.

by ktian1