vGT: validate the tail pointer in the context save/restore
so far we reuse the ring buffer from the guest, and thus expect the
tail pointer unchanged in the time at the context switch. Or else
our save/restore logic races with that.

later we should move to vgt's own ring buffer to avoid that race.
now just warning

by ktian1