vgt: Mark BAR1 as prefetchable
windows driver will check if BAR1 is prefetchable.
the original code hard code BAR1 as 0x4 (64 bit, no prefetchable)
which is not correct. the right way to update BAR is to only
update the bits 4~32, and leave bits 0~3 unchanged, as it is
read-only.

So this patch add correct approach to update BAR, and clean up the
code fo updating BAR. As side effect, the BAR1 is correctly set as
prefetchable.

Signed-off-by: Yu Ke <ke.yu@intel.com>