Fix issue: hvm domain has no display on CRT monitor when display switched back
Because of detection of CRT monitor depends on series steps and need
hardware feedback. This is impossile for hvm domain to
complete this if it can not get real hardware access. Add handler
to emulate this behavior.

And use a flag to indicate whether CRT monitor is being plugged in physically.
This emulation is heavily rely on i915 behavior. This may need further fix.

by axu