uevent fix: fix referencing the address of variable vgt_uevents_bitmap
When delaring bitmap, the variable vgt_uevents_bitmap is actually a
array. It should not add "&" in front of it to extract its address.

Luckily, usually for a general "int a[10]", the value of "a" is equal
to "&a", both equal to the address where array data placed.

Anyway, it will be a serious error if vgt_uevent_bitmap itself is a
pointer.

by axu