Add reg_owner sysfs interface
When debugging going on, more and more regs are involed.
To be clear what kind of type they belong to or which domain
has right to access the hardware. It is important to know
these info dynamically.

This interface is placed at /sys/kernel/vgt/control/reg_owner
When you want to know some register's type and ownership info
you can echo its hexical offset into this interface and use cat
to get the info you want.
e.g. You want to know the info of reg 0xe1180, you can do it
like this:
	echo 0xe1180  > /sys/kernel/vgt/control/reg_owner
	cat /sys/kernel/vgt/control/reg_owner

And the output should be like this:
Reg(000e1180) ownership info:

Type #:              [006]
INVALID:             [no ]
GLOBAL:              [no ]
RCS:                 [no ]
BCS:                 [no ]
VCS:                 [no ]
RENDER:              [no ]
DISPLAY:             [yes]
PM:                  [no ]
MGMT:                [no ]

Owner Domain         [00000000]

The field "Type #" is this type's enum value. The following 9 lines
indicate which type of this reg belongs to. [yes] means it is that
type. Otherwise means not. The field "Owner Domain" means current
which domain has right access the hardware status of this reg. In
the above example, domain 0 has the hardware access grant to this reg.
when this field is "deadbeef" or some other number beyond your expectation.
It means something is wrong.

by axu