EDID virtualization and related stuff
This commit contains the implementation of EDID virtualization, and the features
related to EDID which includes:

1, full virtualization of GPIO MMIO. GPIO is used for I2C bit-banging algorithm.
Windows driver does not use GPIO. In 3.4 or older i915 driver, I2C is used for
EDID reading and SDVO initialization. In 3.5 kernel GMBUS will replace GPIO with
GMBUS. It is safe to drop GPIO support.

In this commit, GPIO implementation by default is disabled. (previously tested
with 3.4 kernel and linux hvm)

2, full virtualization of GMBUS MMIO. GMBUS is more convenient interface
comparing with GPIO. So far from windows and linux driver, the only observed
usage case is for EDID. GMBUS4 for interrupt mask and GMBUS5 for byte index are
not handled in this commit.

3, full virtualization of I2C-over-AUX_CH support. AUX_CH has two kinds of
usage. One is native access and another is I2C-over-AUX_CH. Only the latter
one is fully virtualized.

4, EDID virtualization based on the interface above. The EDID information will
be read once in vgt driver initialization. At the time of vgt instance
initialization, the EDID information will be propagated to vgt data structure.
After that, GMBUS or AUX_CH access for EDID will get the cached data.So far,
CRT, LVDS, DP monitors are tested. Code for HDMI is there but not tesed yet.

5, i915 driver change to disable GPIO access and use GMBUS instead.

6, Hotplug response in terms of EDID. The physical hot plug will generate
uevent to user space (implemented by Eric, Xu). Currently as a reference code,
the user space will write sys fs interface when receiving the uevent. Receiving
the notify, VGT driver then propagate the updated EDID to VM and propagate
virtual interrupt to VMs.