vGT: vgt_display_owner_store: remove display_switched and add locking
This is to fix [VGT BUGID #88]

Currently vgt_display_owner_store() only records next_display_owner and
increases display_switched, and it delays the actual
do_vgt_display_switch() to vgt_thread and it only happens after render
switch starts and the engines are empty(this is a must because
display-releated states can be modified by ringbuffer commands) AND
atomic_read(&display_switched) == 1.

This is buggy, supposing:
1) we create a vgt linux hvm guest and put the guest at grub menu;
2) we try to echo the HVM domain id to the display_owner sys file;
nothig will happen since the guest hasn't enabled ringbuffer at all, but
a new user may feel strange so the user can try do the echo again and
this changes the display_switched to 2.
3) let the grub boot the guest
4) we'll find we can't do any display_owner switch at all!

I made this patch to
1) remove the variable display_switched and use (next_display_owner !=
NULL) as a flag there is a pending request of display_owner switch;
2) after do_vgt_display_switch() is done, reset next_display_owner to NULL;
3) when there is such a pending request, deny any further request by
return -EBUSY;
4) also add proper locking.

Signed-off-by: Dexuan Cui <dexuan.cui@intel.com>