CRT hotplug: move vreg status update from read hook to hotplug irq handler
This is due to that we should update vreg status where the event took place.
The PCH_ADPA read hook is still needed, it is used for dectecting status of
CRT port during boot stage. The detection event took place when writing
PCH_ADPA, but whether or not the CRT monitor sucessfully detected can only
be decided from reading PCH_ADPA after some time. So this kind of logic still
kept in read hook.

by axu