make whitelist match original state
when cleaning up boottime policy, we whitelist all untracked
regs to remove the one-flag-for-all trickiness. However we
progress a bit farther to cause regression. some regs previously
only accessible are now accessible to the owner. This may cause
trouble if current context switch logic doesn't cover them.

with this patch we roll back to the exact same policy as
earlier stage for new added regs, i.e. mark with owner type
if covered in the coarse-grained ranges, or mark as boottime only.

Fix reopened bug [89], that LVDS is dimmed in lowest PWM level,
and bug [101], that brightness control in dom0 doesn't work.

by ktian1