MMIO handler for FORCE_WAKE_ACK on Haswell
The FORCE_WAKE ACK MMIO address on Haswell is different from SNB/IVB. This patch
is to return the wanted value to i915 driver on Haswell. Without the handlere,
on Haswell alpha platform, there will be lots of force wake time out messages in
drm; on Haswell Beta platform, the kernel will hang on some MMIO READ/WRITE.

The reason of hanging on HSW Beta without the patch seems to be hardware related.
Without the patch there would be more READ/WRITE to the hardware MMIO of
FORCE_WAKE_ACK_HSW(ox130044). Seems that it caused problem on HSW beta. Anyway,
problem disappeared after we add the handler.

Signed-off-by: Lv, Zhiyuan <zhiyuan.lv@intel.com>