Be able to build VGT as kernel module other than built-in
Adding needed symbol exporting so that the vGT driver could be build
as a kernel module. Dependence between VGT and AGP_INTEL module was
enforced so that VGT module could be always loaded in advance.
Otherwise, there will be unhandled gp fault in AGP probing since VGT
is not yet inited to register the needed gp fault handler.

Another VGT related module is "XEN_VGT_EMULATOR". In order to build
it as kernel module, more kernel symbols needs to be exported. It is
doable but does not seem to be valuable. So it is still kept as
built-in.

Config file is kept unchanged. Just change "CONFIG_XEN_VGT_I915" to
"m" to build kernel module. In that case, "AGP_INTEL" and "DRM_I915"
will be built as kernel modules as well.

Signed-off-by: Lv, Zhiyuan <zhiyuan.lv@intel.com>

Conflicts:
	drivers/gpu/drm/Kconfig

Conflicts:
	drivers/char/agp/Kconfig
	drivers/xen/events.c