vgt: add klog facility for buck kernel log
there is case that we need to collect large amount of kernel log,
e.g. logging all the guest gfx command in command parser, to better
understand guest behavor.

printk() is not suitable here, as the log msg is too large and frequent.

This patch introduces the klog facility for this. the idea is simple:
klog_printk() will write log to kernel buffer, and transmit the buffer
to userspace via relayfs. userspace app will receive the log, and write
it to local file.

the logging is off by default, thus has no performance impact in normal case.
the userspace app will instrument kernel to turn on the logging via debugfs
when userspace app start. and it will turn logging off when app exit.

This code is customized from relayfs example code:
http://relayfs.sourceforge.net/examples.html

Signed-off-by: Xu Anhua <anhua.xu@intel.com>
Signed-off-by: Yu Ke <ke.yu@intel.com>