vGT: _hvm_mmio_emulation: add the handling for the string-copy case
Also made some space/tab cleanup.

Without this patch, creating win7 guest on beta HSW can hang dom0.

BTW, on the SNB laptop, with the patch, win7 guest (with gfx driver
version:0814) tres to access GM addr not owned by itself and this will
trigger an ASSERT in the vgt driver: this bug 119. To work this
around for now, we do the memcpy only on HSW in the case
(req->dir == IOREQ_READ).

Signed-off-by: Dexuan Cui <dexuan.cui@intel.com>