track saved/restored registers
this is useful to capture risky regs which allows pReg access
but is not saved/restored at approriate context switch time. Now
only a global flag is turned on/off to mark the trace period, within
which all mmio accesses are deemed saved. This is basically fine for
render context switch, because the save/restore process is
straigthforward around an array. But display context switch is more
complex, which may involve registers accessed by vGT driver itself
for mode set which however may not be restored later. Later we
should wrap all restore actions with a single interface. But for
now it's enough to show some basic information.

by ktian1