Break compile time dependence between xen_vgt and i915
This change is to simplify symbol dependence between modules so that the i915
could be built easier in the case that VGT code is not available. Two global
function pointers, corresponding to the two being called by i915, are defined
as call back and will be registered during xen_vgt module initialization.

After this change, the module load sequence of xen_vgt and i915 cannot be
enforced by the caller/callee relationship. In order to make sure that xen_vgt
is loaded before i915 module, a config file vgt.conf is added. It should be
copied to /etc/modprobe.d/ to ensure the sequence when vgt is built as a
module. When vgt is built in, the sequence is enforced by arch_initcall() v.s.
module_init().

Signed-off-by: Zhiyuan, Lv <zhiyuan.lv@intel.com>