Avoid conflict value of VIRQ_VGT_GFX
We did not expect xen hypervisor to send any eventchn of VIRQ_VGT_GFX,
but previously VIRQ_VGT_GFX is defined as 7, which is the same as
"VIRQ_XENOPROF" in xen code. This commit is to change the definition
to use a non-used value.

In future when we rebase xen code to latest version, this definition
should be checked again to avoid any conflicts.

Signed-off-by: Zhiyuan, Lv <zhiyuan.lv@intel.com>