clear bit in started_rings when the ring is disabled
started_rings was introduced to prevent context switch happend
when a VM in console mode. Now with QoS introduced, timer_based_qos did
not use it anymore. But it is still used by shadow_tail_based_qos.
Let's keep it right now: it can be used to identifed the engines
that has been initialized but no commands submitted.

Let's clear the relevant bit when a ring is disabled.

by axu