vgt: remove some sensitive info from the code
The "sensitive" regs can not be found in the i915 code, and can't be found
in the public SNB/IVB specs at http://intellinuxgraphics.org/documentation.html.
For the regs, we remove their macro definitions and use the numerical numbers
instead.

We also rename hdcp related handlers' name to vague names.

This can't be found in public doc:
_REG_LVDS_INT_CTL1

In i915 there are GEN6_BLITTER_ECOSKPD and ECOSKPD, but no VCS_ECOSKPD:
_REG_VCS_ECOSKPD
And let's remove VCS_ECOSKPD from the comment.

This can't be found in public doc:
_REG_GTDRIVER_MAILBOX_DATA1

The *_PSMI macros and _REG_RENDER_PIPE_STATUS are in public spec but the
public spec does not identify these registers by name.

_REGBIT_FDI_TX_PORT_WIDTH_MASK is not defined in public spec or i915 code,
but there is a corresponding definition for RX with same bit pattern in
public spec.

We also fix some license issues and add the missing Copyrights:
The vgt code is GPLv2 only(not dual BSD/GPL license), and when we copy code
from another file, we must copy the Copyright too.

Signed-off-by: Dexuan Cui <dexuan.cui@intel.com>