Use fastpath switch by default
This commit is to change the default switch method to fast path. Right
now the slow path switch only switch VGA (no HDMI and DP yet), and the
VGA part is broken on Haswell.

As earlier discussed, the slow-path display switch has dependence of the
correct MMIO emulation so that we could restore correct Display state.
That part needs some big efforts. Right now, we just turn back to use
fast-path display switch, so that other testings with VGA are not blocked.
The fast-path switch looks working fine in the case that there is no
resolution change.

This commit also has a literal change for the printing log. Fix a format
issue. (losing ")" in message)

Signed-off-by: Zhiyuan Lv <zhiyuan.lv@intel.com>