Remove I2C bit-bang emulation. Physically
The GPIO emulation, in another words, I2C bit-bang emulation, was
removed in this commit. The historical reason to support this is
when we did the EDID virtualization, i915 driver is still using
GPIO for I2C bit-bang to read the EDID information. Bit-wise
emulation is complicated, and it turned out not to be worthwhile.

The GPIO emulation used to work. We ever let both dom0 and linux
hvm read EDID information successfully through GPIO interface. It
was removed just because it is not useful anymore.

Signed-off-by: Zhiyuan Lv <zhiyuan.lv@intel.com>