Bug fix of #206, the win7 failure in second booting
This commit is the bug fix of below:
http://vmm-qa.sh.intel.com/vgt/bugzilla/show_bug.cgi?id=206

The root cause of this bug is that, there are GMBUS3 read operations
which are trying to access some EDID information that is not
available. In normal case, the read should not happen due to the
earlier failure of GMBUS state check. But the current code somehow
does not let gfx driver know that the EDID information is not
available.

The bug was introduced by below commit:
bec2f53ccf9ae5b940705613270f813a95b8074e

Previously, for the port which does not have EDID information
available, the code will not return GMBUS_ACTIVE bit in GMBUS2
reading, hence gfx driver is aware of the failure. Above commit
changed the code by using GMBUS_NAK bit in GMBUS2. The idea should be:
in the case of no EDID information, there will be no response from
GMBUS status read. However, the GMBUS_NAK bit will be cleared in the
GMBUS reset, then it is broken.

The fix is to set the GMBUS_NAK bit during GMBUS reset.

Signed-off-by: Zhiyuan Lv <zhiyuan.lv@intel.com>