Remove next_display_owner
After discussion with Anhua and Dexuan, seems that the global variable
"next_display_owner" was introduced by historical reasion and is not
needed anymore.

In the original code, "next_display_owner" is used to do some kind of
protection so that there will not be two "display switch" happening
concurrently. It was to resolve some issues when
"do_vgt_display_switch" was also called in vgt_thread. But now, it has
not been the case. And, there is a lock around "do_vgt_display_switch"
which can also prevent the concurrent switches.

Signed-off-by: Zhiyuan Lv <zhiyuan.lv@intel.com>