introduce foreground_vm
Control whether a vm is running in foreground, that is, whether it is
visible on screen. A vm does not necessarily become display owner in
order to show up on screen. Just set the foreground_vm in sysfs
interface.

Meanwhile, the setting of display_owner was disabled, which means only
dom0 could be the display owner to touch display hardware. We will
keep this until we have a complete slow-path display switch support.

Signed-off-by: Zhiyuan Lv <zhiyuan.lv@intel.com>