Change interrupt interval from 500ms to 16ms
This is suggested by Kevin. Originally for the VM which is not running
with display owner, it will get gfx interrupts in very slow frequency.
(500ms, 2fps). This change is to let non-display-owned VM be able to
reach 60FPS also (16ms interval).

Kevin is changing interrupt framework. That code is stay tuned.

Signed-off-by: Zhiyuan Lv <zhiyuan.lv@intel.com>