Display MMIOs audit for full virtualization
All the tracked MMIOs were went through. For those that are Display
related, if they were previously defined as WA, they are changed to
DPY this time.

As a result, when a VM is not display owner, it cannot touch
display hardware through these MMIOs. It makes hvm much safer to boot
and show up in the mode that it is not display owner.

Signed-off-by: Zhiyuan Lv <zhiyuan.lv@intel.com>