Make DE_RRMR to be WA type
The MMIO DE_RRMR's type was changed from F_DPY to be F_WA. Dexuan and
Kevin did the investigation of bug 250:

http://vmm-qa.sh.intel.com/vgt/bugzilla/show_bug.cgi?id=250
"tests in 3dmark can not run"

Found that the reason is the DPY register DE_RRMR impacted render
ring. Render ring waited for an event but that was masked due to the
incorrect value in DE_RRMR.

The error happens when dom0 is display_owner, and set DE_RRMR to be
some value, say, FFFFFFFF, then hvm boots without display_owner. In
original policy, DE_RRMR is fully virtualized and cannot change
hardware. Then render ring hangs.

Right now, the fix is to let hvm be able to change hardware for RRMR.
The policy of RRMR needs to be carefully considered since it is quite
special. It belongs to DPY but impacts rendering. We could handle it
like render MMIO, but if its value does not change frequently, we need
to consider whether it is worthwhile.

Signed-off-by: Zhiyuan Lv <zhiyuan.lv@intel.com>