Add PIPE C support, and some MMIO correction
This commit adds the missing PIPE C related MMIO entries in the MMIO
handler table. Missing those will cause XenGT to report "Untracked MMIO"
and the PIPE C cannot work functionaly.

This commit also correct some MMIOs that are different in SNB/IVB and HSW.
One example is the CURBXXX MMIOs. In IVB/HSW, the MMIOs are still there
but the address changed.

Signed-off-by: Zhiyuan Lv <zhiyuan.lv@intel.com>