Add handler for CURSOR and SPRITE SURFLIVE support
They are MMIOs like DSPSURFLIVE MMIOs, which are readonly and
automatically updated when base address change happens. Sprite
SURF MMIOs were added in this commit also.

The Cursor SURFLIVE registers are not defined in i915_reg.h, but
they could be found in IVB PRM (A/B/C) and SNB PRM (A/B). So they
are not sensitive registers.

Signed-off-by: Zhiyuan Lv <zhiyuan.lv@intel.com>