vgt: fix HVM's BAR virtualization
This is to fix bug 262:
http://vmm-qa.sh.intel.com/vgt/bugzilla/show_bug.cgi?id=262

HVM's vBAR's initial value is copied from dom0's(i.e., the pBAR's
base), so when hvmloader tries to get the bar size by
            bar_data = pci_readl(devfn, bar_reg);
            pci_writel(devfn, bar_reg, ~0);
            bar_sz = pci_readl(devfn, bar_reg);
            pci_writel(devfn, bar_reg, bar_data);
in vgt_hvm_map_apperture(), vgt maps the vBAR=pBAR(the base is 0xa0000000
on the B0 mobile HSW, and the size is 512MB) on the second pci_writel().

Later when hvmloader assigns a real vBAR(i.e., 0xc0000000), vgt unmaps
vBAR 0xa0000000, and maps 0xc0000000 to pBAR.
Unluckily the unmapping here clears [0xa0000000, 0xa0000000+512MB) from
HVM's P2M table, but actually this ranges can be E820_RAM!

When HVM tries to access an ordinary RAM addr 0xa45fffe8 for the first
time, Xen wrongly treats this addr as MMIO and forwards it to QEMU.

QEMU can't handle this since it finds the addr is RAM but it can't map
the GPFN properly since the related P2M mapping is invalid. As a result,
QEMU will crash after receiving a signal SIGBUS.

After a discussion, the correct fix should be:
1) the vCOMMAND reg's initial value of HVM should clear the related bits for
PIO/MMIO.
2) when vPIO/VMMIO are disabled, vgt shouldn't update the vBAR/pBAR mapping.
3) when vPIO/VMMIO are enabled or being enabled, vgt should update the
mapping when vBARs or vCOMMAND are written.

Signed-off-by: Dexuan Cui <dexuan.cui@intel.com>