Fix DP port detection for EDID/DPCD
The original code uses DP_AUX_CH_CTL to check whether a DP port is
enabled, but actually DP_CTL should be used instead (for DPA, and
SFUSE_STRAP for others). The original code always get "true" for the
display check, and then will get all zero EDID and DPCD for the ports
without monitor connected.

And even if we use that, We still cannot make sure there is a DP/eDP
monitor connected. But anyway, the port check is needed.

For DP, we should read DPCD and hot plug detection bit for the
connection check.

Signed-off-by: Zhiyuan Lv <zhiyuan.lv@intel.com>