Fix some timing issues for AUX_CH on eDP
Some timing config for AUX_CH is very device specific. This commit
is to add the support of eDP on HSW ULT. More decent support needs
to check many more device types and set different value.

However, I do not want this part of code to be like i915 driver.
Later this part of code will be cleaned up further and reconsider
the way we cache EDID and DPCD information.

Right now, live with it and this commit makes hvm working with eDP
on HSW ULT.

Signed-off-by: Zhiyuan Lv <zhiyuan.lv@intel.com>