Display related option changes
This commit did below display related option changes:

1, introduce hvm_boot_foreground for hvm booting. If this option
is used, hvm will be the current_foreground_vm after it is created,
so it could be visible from boot time. Previously we borrowed
"hvm_dpy_owner" for this, but it is misleading. The original
usage of this option is to set hvm to be the display owner,
which is different from what we want.

2, introduce hvm_display_owner to be a replacement of hvm_dpy_owner.
This option will make the hvm to be the display owner after a hvm
is created. It is mainly for debugging purpose, and will be set
true if users use "hvm_super_owner" option.

3, Deprecate "hvm_dpy_owner". Since we ever used this option for
different purpose, (set current foreground vm or set display owner vm),
in order not to confuse users further, this option is deprecated.
People should use "hvm_display_owner" or "hvm_boot_foreground".

Signed-off-by: Zhiyuan Lv <zhiyuan.lv@intel.com>