Change surflive virtualization policy, fix GPU reset issue
This commit changes our virtualization policy for surflive
registers. Previously we do the surflive vreg/sreg update when
surf registers are written in background mode. That could
introduce some synchronization issue. For instance, SURF registers
are written when VM is running in foreground mode, then VM is
switched to background mode. After that, gfx driver may read
surflive to check whether the SURF registers have been written
into real hardware. But what driver can get is the out-of-date
vreg in surflive, and then the check will fail. It is convinced
to be the root cause of the GPU reset when we do display switch
with 3D games running inside win7.

The new virtualization policy is pretty simple and clear. We fully
virtualize SURFLIVE registers. When there is MMIO read of SURFLIVE,
vgt will return the SURF register vreg value.

This does not accurately emulate GPU hardware behavior because it
may update surflive value too early. (in real hardware, the SURFLIVE
can only be updated in the next VBLANK after SURF register write).
But our policy is pretty safe. And solved problem.

Signed-off-by: Zhiyuan Lv <zhiyuan.lv@intel.com>