vgt: remove the unnecessary GPLv2 headers from dual GPLv2/MIT licensed files
By default a code file is GPLv2 in Linux kernel, so no need to have the GPLv2
headers for dual GPLv2/MIT licensed files, just as i915 driver does.

NOTE:
Also removed the MIT header from include/xen/interface/hvm/ioreq.h:
The file was copied from upstream Xen(it only has an MIT headre there. But
since Xen itself is GPL, this means the file is under dual GPL/MIT in Xen),
and it should be OK for us to use it as GPLv2 only in the vGT project.

Signed-off-by: Dexuan Cui <dexuan.cui@intel.com>