Introduce per-vm's "presented_ports"
This commit introduces per-vm's presented_ports, and renamed original
global variable in pdev called "port_detect_status" to be
"detected_ports". In system initialization, our VGT module
will set "detected_ports" based on whether EDID is available.
Then the information will be propagated to "presented_ports"
while propagating EDID.

The "detected_ports" and "presented_ports" could be considered as a
flag of whether EDID is available, that is, whether a monitor is
connected.

Signed-off-by: Zhiyuan Lv <zhiyuan.lv@intel.com>