Validate DPCD information during probing
We always meet the case that AUX_CH read just succeeded (no timeout etc)
but the value in DATA register is just zero. This commit added the
check of DPCD header after the probing. If the "REV" field is zero,
that implies DP monitor is not connected. This kind of check is found
in i915 driver for DP check. We borrowed the logic to here.

So we will do DPCD probing first. If DPCD not available, we do not
need to probe EDID for the port then.

Signed-off-by: Zhiyuan Lv <zhiyuan.lv@intel.com>