Enhance the hotplug handling for CRT
This commit enhanced the hotplug handling for CRT as follow:

1, In CRT hotplug virtual handler, check whether the hotplug is
enabled. If not enabled, do not propagate the interrupt bits and
do not inject the interrupt to VM.

2, Fix the usage of "detected_ports". Previously the value
will be updated in the MMIO read of ADPA, according to the stats
bits. But that may not be correct. The status bits can only be
updated in FORCE_TRIGGER or hot plug interrupt. But the MMIO
reading does not necessarily happen in that situation. Then the
update may be incorrect.

The bit set in crt interrupt physical handler was also removed.
The status will be set when dom0 does the EDID probing before
propagating hotplug into VMs.

3, Change the virtualization policy of FORCE_TRIGGER. We will
always fully virtualize this behavior. The "FORCE_TRIGGER" bit
will not be set to real hardware. Instead, when we detect the
write of this bit from VM (including dom0), we will update
the two status bits according to the monitor connection information.
That is more clear.

Signed-off-by: Zhiyuan Lv <zhiyuan.lv@intel.com>