Hotplug interface fix
Change hotplug interface as follow:

1, In the monitor type definition field, remove the previous DP_B, HDMI_B,
instead, add the PORT_B etc. The reason to do the change is that DP and
HDMI physically share the same ports. The hotplug interrupt itself does
not know whether it is from DP monitor or HDMI monitor. The only to know
this is to check the EDID information. So we simply use ports for the
uevent. vgt module could later know the monitor type.

2, hotplug for SDVO was also removed since it does not seem to be supported
in HSW.

3, Change the "sscanf" format parameter to "%i" so that people could
input hexadecimal data as well as decimal data.

Signed-off-by: Zhiyuan Lv <zhiyuan.lv@intel.com>