Make DP and HDMI who are on the same port equivalent
When we do EDID operations like probling, propagating, deleting,
treat DP and HDMI together. For instance, if programs requre the
propagating of DP_B, we propagate HDMI_B also.

The reason to do this is that DP and HDMI share the same port. In
the current hotplug process, VGT need to probe EDID on a specific
port when physical hotplug interrupt happens (in uevent handler
currently). The only way to know whether a DP or HDMI monitor is
connected to the port is to proble both the two. The same to the
EDID clear and propagation.

Signed-off-by: Zhiyuan Lv <zhiyuan.lv@intel.com>