Add more render related registers
Add more render related registers in the MMIO handler table, although
currently we did not find the access from both linux and widows hvm.
The reason to add them is to let vgt driver to cache the hardware value
for them in system initialization. Right now we only do so for tracked
MMIOs.

Because those MMIOs are candidate for render context switch, it is
necessary to have correct init value in VM's vreg/sreg.

Signed-off-by: Zhiyuan Lv <zhiyuan.lv@intel.com>