Set offset before probing EDID from DP
The previous code did not set the offset 0 before the probing of
EDID from DP. That caused issues on some DP monitors. For DP with
extended blocks, we saw that the second block was read out instead
of the first one.

The fix is to set offset right before EDID reading, referring to the
EDID read operations in i915 code.

Signed-off-by: Zhiyuan Lv <zhiyuan.lv@intel.com>