Enhance the EDID probing for DP
It was spotted that reading EDID from DP sometimes does not work.
This commit enhanced the DP AUX_CH reading according to i915 code
and spec. (This part of code is not related to virtualization)

1, Set clock frequency according to different chip type. Right now
we only handle HSW platform.

2, Try more if AUX_CH transaction fails sometimes. Maximum try 5
times.

With this commit and the previous one, bug #284 should have been
fixed:

"#284: can not collect edid in guest os with eDP and DP port"

DP monitor has been verified to work. eDP not tested due to lack of
platform in Beijing.

Signed-off-by: Zhiyuan Lv <zhiyuan.lv@intel.com>