Change vgt_thread and vgt_emulation_thread as "INTERRUPTIBLE"
The wait_event used in both kernel threads will mark them as
uninterruptible task in waitqueue. When "Detect Hung Task" is
enabled in dom0, if there are no events need to handle for these
threads for more than 120 seconds. Such information will be
printed and also a dumpstack for the thread (if log level is
permitted)

[  250.800795] INFO: task vgt_thread:755 blocked for more than 120 seconds.
[  250.808618] "echo 0 > /proc/sys/kernel/hung_task_timeout_secs" disables this
message

This patch can fix bug#292

Signed-off-by: Anhua Xu <anhua.xu@intel.com>