Fix incorrect padding in struct vgt_if
The the value 0x60 is not correct. It is the byte size for the padding.
When using "uint_32" as the unit, it should use 4 to divide it.

The balloon information(avail_rs) is just located prior this padding
So both Windows and Linux HVM works fine. But the data structure of
minimum required resource (filled by HVM guest) is located after
this padding. So when Windows HVM write its request into PVINFO page,
vgt cannot get the correct value because of different offset used
by each of them.

Signed-off-by: Anhua Xu <anhua.xu@intel.com>