Ignore checking ring idle for those disabled ring
Sometimes, two VM can have different enabled rings.
Like in Linux HVM, RCS,BCS,VCS are enabled; while in
Windows 7 HVM, RCS,BCS,VCS and VECS are enabled. In
such circumstances, when context switching from Linux
HVM to Windows HVM, we should not check if VECS ring
is idle, it is not used by Linux HVM (but it is used
by Windows HVM, so saving context can be ignored but
restoring context is needed), and it is  possible
introduce unknown issues.

This patch can fix bug# 305.

Signed-off-by: Anhua Xu <anhua.xu@intel.com>