Always restore ring MMIO even if the next owner doesn't enable it
In the original logic, Ring MMIO will not be restored to HW when
the next owner does not enable this ring. This might cause the
the below issue (bug# 305):

When dom0 did not enable VECS ring (head/tail will be 0) while
Windows 7 guest enable VECS ring. Assume one scenario like this:
Currently Windows 7 guest is the render owner, after it used up
its time slice, both head/tail of VECS ring will be stopped
at 0xa00600; After saving context of Windows 7 guest, Dom0's context
will be resotred, but its VECS ring MMIO(head/tail) will not be
restored by vgt_restore_ringbuffer(). Because in the logic of
vgt_restore_ringbuffer(), when a ring is disabled in a VM, its
head/tail value will not be restored to HW by vgt_restore_ringbuffer().
And ctl will written as 0 to HW.

So physical head/tail are still 0xa00600. Then vgt_kick_ringbuffers(next);
will be invoked. It will write the next owner's tail value into HW.
In this case, it is dom0's tail whose value is 0. So at this moment,
physical head is still 0xa00600, while tail and ctl become 0.

After dom0 used up its time slice, it will try to stop all 4 rings.
When checking went to VECS, it's ctl is 0, so ring empty check will be
ignored (disabled ring was thought as empty ring). But idle bit (bit 8)
is 0 in 0x8008. Such fact tells that VECS ring never become idle. This
is the error log of bug#305 showed.

Since physically VECS's head is 0xa00600 and tail is 0, physical head not
equal to tail could be root cause that VECS ring was not in idle state.

So this patch try to restore all ring mmio for each ring no matter if it
is enabled by the next owner.

Signed-off-by: Anhua Xu <anhua.xu@intel.com>