Keep the original logic of ring_is_idle()
The reason is vgt could consider a ring as idle because
the enable/disable bit in ctl register is 0. But actually,
in hw the idle bit might not indicate idle state. Right now,
we should strictly follow hw status when conducting context
switch. This is also help to expose potential issue (just
like bug #305)

Signed-off-by: Anhua Xu <anhua.xu@intel.com>