Initial implementation of framebuffer change notifier
Use notify chain to send out notification to other modules in kernel when frame
buffer related attributes have changed. Notify chain was used. The
implementation includes:

1, Define notification event type including FB_MODE_SET_START, FB_MODE_SET_END and
FB_DISPLAY_FLIP. The first two depends on the windows driver's notification through
pv info. So we currently will not generate this.

2, Notification in cmd_parser for display flip.

3, Notification in MMIO handlers for display flip.

4, Notification could have a data attached. Currently we simply provide vm_id
and pipe_id information for a given notification event. The final design of
this depends on requirements.

5, Interface of register or unregister a notifier for other modules.

Signed-off-by: Zhiyuan Lv <zhiyuan.lv@intel.com>