Not to clear CRT status bits when writing 1 to them
Bspec said that CRT status bits in E1100h will be cleard by software
by writing 1 to them. But that is not true in hardware. After asking
hw architect, we got confirmed that it is a BSpec bug. So the handler
was changed accordingly to be consistent with hardware.

The hw behavior was originally noticed by Anhua. It did not cause
problem until we virtualize those bits based on BSpec. Wingfx driver
expects those bits not to be cleard. Otherwise, CRT monitor cannot
be detected in multiple monitor scenario. Xiaowei debugged wingfx
driver and root caused the problem.

Signed-off-by: Zhiyuan Lv <zhiyuan.lv@intel.com>