Fix the EDID reading bug on eDP; #284
This should be a bug introduced by the last code change of AUX_CH
operations for reading hardware EDID. The problem is that the MOT
bit was not set for AUX_CH read and write. That caused the problem
that the stream reading just finished after the first byte reading.
Surprisingly, EDID read for DP, and DPCD read for DP and eDP are
both working even without the fixing. Only EDID read for eDP
encountered problem.

Related bug entry is #284

Signed-off-by: Zhiyuan Lv <zhiyuan.lv@intel.com>