Probe DPCD also when hotplug happens
To update DP monitor's status. Also dpcd should be propagated in
hotplug handling.

Meanwhile, this commit fixed a mismatch ID check for dpcd. The
mismatch will cause the DPCD information not correctly propagated to
hvm in hotplug cases. It happens only in hotplug cases. In other
cases, the DPCD propagation was done with parameter "-1", which means
to propagate all ports, and the mismatch is not an issue. The failure
of propagation may cause hvm to fail on DP monitor detection.

Signed-off-by: Zhiyuan Lv <zhiyuan.lv@intel.com>