Update SDEISR while connected monitors changed
Update the hotplug bits of port B/C/D in SDEISR while monitor connection
changes. The update is based on the information of presented monitor of a hvm.
That update happens in two places: one is in hvm creation, another is in
hotplug event.

The Reasons to do that change is: In current wingfx driver code of
checking monitor attached information, for HDMI or DP on B/C/D port,
the function will first check ISR to determin whether the port is in
live state. We need to give the information to hvm. Otherwise the
monitor check will fail.

For current wingfx code, only SDEISR need such handling. DE interrupt
control register has similar bit for DP A hotplug. But DP A detection
in wingfx code does not use this.

Also a small clean up for SDEISR: _REGBIT_SDVO_B_HOTPLUG, which is
incorrect at least on HSW, and it is not used anyway.

Signed-off-by: Zhiyuan Lv <zhiyuan.lv@intel.com>