Explictly only support HSW platform in XenGT
Initiate NULL context now is only supported on HSW.
So SNB/IVB will fail for this. This patch is used
to fix bug#341.

And after falure of vgt_initialize_pgt_device(),
we cannot "goto err;" and call vgt_destroy(). Because
it will cause NULL pointer when list_del(&pdev->list).
pdev->list is not initialized yet.

This patch is a quick fix for running XenGT on unsupported
platform like IVB/SNB. But this issue reveals that we lack
of handling errors from other parts of vgt_initialize().

These issues cannot be reproduced easlily since most of the
time Dom0 with XenGT can boot up successfully.

Signed-off-by: Anhua Xu <anhua.xu@intel.com>